User permissions

+++ app/Http/Controllers/MenuController.php
@@ -6,6 +6,7 @@
 	use Illuminate\Http\Request;
 	use Illuminate\Support\Facades\Redirect;
 	use App\Http\Requests\Menu\MenuRequest;
+	use Auth;
 
 	class MenuController extends Controller 
 	{
@@ -18,9 +19,13 @@
 
 		public function index()
 		{
-			$allMenuItemsEdit = $this->menuRepo->getAllMenuItems();
+			if (Auth::user()->hasPermission(PermissionsController::PERMISSION_MENU))
+			{
+				$allMenuItemsEdit = $this->menuRepo->getAllMenuItems();
 
-			return view('menu.index', compact('allMenuItemsEdit'));
+				return view('menu.index', compact('allMenuItemsEdit'));
+			}
+			return view('errors.403');
 		}
 
 		/**
@@ -30,9 +35,13 @@
 		 */
 		public function create()
 		{
-			$menuItem = new Menu();
+			if (Auth::user()->hasPermission(PermissionsController::PERMISSION_MENU))
+			{
+				$menuItem = new Menu();
 
-			return view('menu.create', compact('menuItem'));
+				return view('menu.create', compact('menuItem'));
+			}
+			return view('errors.403');
 		}
 
 		/**
@@ -42,9 +51,13 @@
 		 */
 		public function store(MenuRequest $request)
 		{
-			$this->menuRepo->create($request->all());
+			if (Auth::user()->hasPermission(PermissionsController::PERMISSION_MENU))
+			{
+				$this->menuRepo->create($request->all());
 
-			return Redirect::route ('menu.index');
+				return Redirect::route ('menu.index');
+			}
+			return view('errors.403');
 		}
 
 		/**
@@ -56,9 +69,13 @@
 		 */
 		public function edit($id)
 		{
-			$menuItem = $this->menuRepo->get($id);
+			if (Auth::user()->hasPermission(PermissionsController::PERMISSION_MENU))
+			{
+				$menuItem = $this->menuRepo->get($id);
 
-			return view('menu.edit', compact('menuItem'));
+				return view('menu.edit', compact('menuItem'));
+			}
+			return view('errors.403');
 		}
 
 		/**
@@ -70,58 +87,58 @@
 		 */
 		public function update(MenuRequest $request)
 		{
-		   $item = $this->menuRepo->get($request->id);
-		   $item->name = $request->name;
-		   $item->link = $request->link;
-		   $item->publish = $request->publish;
-		   $this->menuRepo->update($item);
+			if (Auth::user()->hasPermission(PermissionsController::PERMISSION_MENU))
+			{
+				$item = $this->menuRepo->get($request->id);
+				$item->name = $request->name;
+				$item->link = $request->link;
+				$item->publish = $request->publish;
+				$this->menuRepo->update($item);
 
-			return Redirect::route('menu.index');
+				return Redirect::route('menu.index');
+			}
+			return view('errors.403');
 		}
 
 		public function updateMenuOrder(Request $request)
 		{
-			$parentId = null;
-			$array = [];
-			$allMenuItems = $this->menuRepo->getAll();
-
-			foreach($request->all() as $key => $requestItem) //loop trough the names of the textfields
+			if (Auth::user()->hasPermission(PermissionsController::PERMISSION_MENU))
 			{
-				if (! is_string( $key ))
+				$parentId = null;
+				$array = [];
+				$allMenuItems = $this->menuRepo->getAll();
+
+				foreach ($request->all() as $key => $requestItem) //loop trough the names of the textfields
 				{
-					foreach($allMenuItems as $oldIndex => $oldItem)
-					{
-						if ($oldItem->menuId == $key)
-						{
-							unset($allMenuItems[$oldIndex]);
+					if (!is_string($key)) {
+						foreach ($allMenuItems as $oldIndex => $oldItem) {
+							if ($oldItem->menuId == $key) {
+								unset($allMenuItems[$oldIndex]);
 
-							break;
+								break;
+							}
 						}
-					}
-					$requestItemPart = explode(".", $requestItem);
+						$requestItemPart = explode(".", $requestItem);
 
-					if ($requestItemPart[0] == 0)
-					{
-						$array = [];
-						$this->menuRepo->updateMenuItemOrder($key, $requestItemPart[1], NULL );
-					}
-					else
-					{
-						if(isset($array[$requestItemPart[0]]) || empty($array[$requestItemPart[0]]))
-						{
-							$array = array_add($array, $requestItemPart[0], $parentId);
+						if ($requestItemPart[0] == 0) {
+							$array = [];
+							$this->menuRepo->updateMenuItemOrder($key, $requestItemPart[1], NULL);
+						} else {
+							if (isset($array[$requestItemPart[0]]) || empty($array[$requestItemPart[0]])) {
+								$array = array_add($array, $requestItemPart[0], $parentId);
+							}
+							$this->menuRepo->updateMenuItemOrder($key, $requestItemPart[1], $array[$requestItemPart[0]]);
 						}
-						$this->menuRepo->updateMenuItemOrder($key, $requestItemPart[1], $array[$requestItemPart[0]]);
+						$parentId = $key;
 					}
-					$parentId = $key;
 				}
-			}
 
-			foreach($allMenuItems as $oldItem)
-			{
-				$this->menuRepo->destroy($oldItem->menuId);
-			}
+				foreach ($allMenuItems as $oldItem) {
+					$this->menuRepo->destroy($oldItem->menuId);
+				}
 
-			return Redirect::route('menu.index');
+				return Redirect::route('menu.index');
+			}
+			return view('errors.403');
 		}
 	}
